feat: Update col type in db

+++ db/migrations/000001_init_schema.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "accounts" (
     "id"            bigserial       PRIMARY KEY,
     "owner"         varchar         NOT NULL,
-    "balance"       numeric(10,2)   NOT NULL,
+    "balance"       bigint          NOT NULL,
     "currency"      varchar         NOT NULL,
     "created_at"    timestamptz     NOT NULL DEFAULT (now())
 );
@@ -9,7 +9,7 @@ CREATE TABLE "accounts" (
 CREATE TABLE "entries" (
     "id"            bigserial       PRIMARY KEY,
     "account_id"    bigint          NOT NULL,
-    "amount"        numeric(10,2)   NOT NULL,
+    "amount"        bigint          NOT NULL,
     "created_at"    timestamptz     NOT NULL DEFAULT (now())
 );
 
@@ -17,7 +17,7 @@ CREATE TABLE "transfers" (
     "id"                bigserial       PRIMARY KEY,
     "from_account_id"   bigint          NOT NULL,
     "to_account_id"     bigint          NOT NULL,
-    "amount"            numeric(10,2)   NOT NULL,
+    "amount"            bigint          NOT NULL,
     "created_at"        timestamptz     NOT NULL DEFAULT (now())
 );
